working on charts

+++ app/_lib/utils/months.ts
@@ -1,4 +1,4 @@
 const months = Array.from({ length: 12 }, (_, i) =>
-  new Date(2024, i, 1).toLocaleString("en-US", { month: "long" })
+  new Date(2024, i, 1).toLocaleString("en-US", { month: "short" })
 );
 export default months;
